Refactor and split into multiple files
Each module (and submodules) has no its own file.
Comments were also refactored to markdown comments for
the docs.
