+++ api/rust/suborbital/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "suborbital"
 version = "0.12.0"
-authors = ["cohix <connorjhicks@gmail.com>"]
+authors = ["cohix <connorjhicks@gmail.com>", "Florian Fromm <flrn.frmm@gmail.com>"]
 edition = "2018"
 description = "Suborbital Wasm Runnable API"
 license = "Apache-2.0"
